+++ processor/src/main/java/arez/processor/ComponentDescriptor.java
@@ -58,7 +58,6 @@ final class ComponentDescriptor
   private static final Pattern OBSERVABLE_REF_PATTERN = Pattern.compile( "^get([A-Z].*)ObservableValue$" );
   private static final Pattern COMPUTABLE_VALUE_REF_PATTERN = Pattern.compile( "^get([A-Z].*)ComputableValue$" );
   private static final Pattern OBSERVER_REF_PATTERN = Pattern.compile( "^get([A-Z].*)Observer$" );
-  private static final Pattern PRIORITY_OVERRIDE_PATTERN = Pattern.compile( "^(.*)Priority$" );
   private static final Pattern SETTER_PATTERN = Pattern.compile( "^set([A-Z].*)$" );
   private static final Pattern GETTER_PATTERN = Pattern.compile( "^get([A-Z].*)$" );
   private static final Pattern ID_GETTER_PATTERN = Pattern.compile( "^get([A-Z].*)Id$" );
@@ -163,7 +162,6 @@ final class ComponentDescriptor
   private final Map<String, InverseDescriptor> _inverses = new LinkedHashMap<>();
   private final Collection<InverseDescriptor> _roInverses =
     Collections.unmodifiableCollection( _inverses.values() );
-  private final Map<String, CandidateMethod> _priorityOverrides = new LinkedHashMap<>();
 
   ComponentDescriptor( @Nonnull final ProcessingEnvironment processingEnv,
                        @Nonnull final String type,
@@ -1343,8 +1341,6 @@ final class ComponentDescriptor
 
     autodetectObservableInitializers();
 
-    linkPriorityOverrideMethods();
-
     /*
      * All of the maps will have called remove() for all matching candidates.
      * Thus any left are the non-arez methods.
@@ -2339,34 +2335,6 @@ final class ComponentDescriptor
     }
   }
 
-  private void linkPriorityOverrideMethods()
-    throws ProcessorException
-  {
-    for ( final Map.Entry<String, CandidateMethod> entry : _priorityOverrides.entrySet() )
-    {
-      final String name = entry.getKey();
-      final CandidateMethod method = entry.getValue();
-      final MemoizeDescriptor memoize = _memoizes.get( name );
-      if ( null != memoize )
-      {
-        memoize.setPriorityOverride( method );
-      }
-      else
-      {
-        final ObserveDescriptor observe = _observes.get( name );
-        if ( null != observe )
-        {
-          observe.setPriorityOverride( method );
-        }
-        else
-        {
-          throw new ProcessorException( "@PriorityOverride target has no corresponding @Memoize or " +
-                                        "@Observe methods", method.getMethod() );
-        }
-      }
-    }
-  }
-
   private boolean analyzeMethod( @Nonnull final ExecutableElement method,
                                  @Nonnull final ExecutableType methodType )
     throws ProcessorException
@@ -2401,8 +2369,6 @@ final class ComponentDescriptor
       AnnotationsUtil.findAnnotationByType( method, Constants.COMPONENT_NAME_REF_ANNOTATION_CLASSNAME );
     final AnnotationMirror postConstruct =
       AnnotationsUtil.findAnnotationByType( method, Constants.POST_CONSTRUCT_ANNOTATION_CLASSNAME );
-    final AnnotationMirror priorityOverride =
-      AnnotationsUtil.findAnnotationByType( method, Constants.PRIORITY_OVERRIDE_ANNOTATION_CLASSNAME );
     final AnnotationMirror ejbPostConstruct =
       AnnotationsUtil.findAnnotationByType( method, Constants.EJB_POST_CONSTRUCT_ANNOTATION_CLASSNAME );
     final AnnotationMirror preDispose =
@@ -2532,11 +2498,6 @@ final class ComponentDescriptor
       setComponentTypeNameRef( method );
       return true;
     }
-    else if ( null != priorityOverride )
-    {
-      addPriorityOverride( priorityOverride, method, methodType );
-      return true;
-    }
     else if ( null != ejbPostConstruct )
     {
       throw new ProcessorException( "@" + Constants.EJB_POST_CONSTRUCT_ANNOTATION_CLASSNAME + " annotation " +
@@ -2592,66 +2553,6 @@ final class ComponentDescriptor
     else
     {
       return false;
-    }
-  }
-
-  private void addPriorityOverride( @Nonnull final AnnotationMirror annotation,
-                                    @Nonnull final ExecutableElement method,
-                                    @Nonnull final ExecutableType methodType )
-  {
-    MemberChecks.mustNotBeAbstract( Constants.PRIORITY_OVERRIDE_ANNOTATION_CLASSNAME, method );
-    MemberChecks.mustBeSubclassCallable( getElement(),
-                                         Constants.COMPONENT_ANNOTATION_CLASSNAME,
-                                         Constants.PRIORITY_OVERRIDE_ANNOTATION_CLASSNAME,
-                                         method );
-    MemberChecks.mustNotThrowAnyExceptions( Constants.PRIORITY_OVERRIDE_ANNOTATION_CLASSNAME, method );
-    MemberChecks.mustReturnAValue( Constants.PRIORITY_OVERRIDE_ANNOTATION_CLASSNAME, method );
-
-    final List<? extends VariableElement> parameters = method.getParameters();
-    if ( !( parameters.isEmpty() || 1 == parameters.size() && parameters.get( 0 ).asType().getKind() == TypeKind.INT ) )
-    {
-      throw new ProcessorException( "@PriorityOverride target must have no parameters or a " +
-                                    "single int parameter", method );
-    }
-
-    final TypeMirror type = method.getReturnType();
-    if ( TypeKind.INT != type.getKind() )
-    {
-      throw new ProcessorException( "@PriorityOverride target must return an int value", method );
-    }
-
-    final String name = derivePriorityOverrideName( method, annotation );
-    _priorityOverrides.put( name, new CandidateMethod( method, methodType ) );
-  }
-
-  @Nonnull
-  private String derivePriorityOverrideName( @Nonnull final ExecutableElement method,
-                                             @Nonnull final AnnotationMirror annotation )
-  {
-    final String declaredName = getAnnotationParameter( annotation, "name" );
-    if ( Constants.SENTINEL.equals( declaredName ) )
-    {
-      final String name = ProcessorUtil.deriveName( method, PRIORITY_OVERRIDE_PATTERN, declaredName );
-      if ( null == name )
-      {
-        throw new ProcessorException( "Method annotated with @PriorityOverride should specify name or be " +
-                                      "named according to the convention [name]Priority", method );
-      }
-      return name;
-    }
-    else
-    {
-      if ( !SourceVersion.isIdentifier( declaredName ) )
-      {
-        throw new ProcessorException( "@PriorityOverride target specified an invalid name '" + declaredName +
-                                      "'. The name must be a valid java identifier.", method );
-      }
-      else if ( SourceVersion.isKeyword( declaredName ) )
-      {
-        throw new ProcessorException( "@PriorityOverride target specified an invalid name '" + declaredName +
-                                      "'. The name must not be a java keyword.", method );
-      }
-      return declaredName;
     }
   }